moving the search component out of the members folder
