+++ app/components/search_component.html.erb
