+++ app/views/groups/members/index.turbo_stream.erb
@@ -1,7 +1,7 @@
 <% if @tab != "invited_groups" %>
 
   <%= turbo_stream.update "members_search_bar" do %>
-    <%= render Members::SearchComponent.new(
+    <%= render SearchComponent.new(
       @q,
       @tab,
       group_members_url(**request.query_parameters),
