+++ test/components/members/search_component_test.rb
@@ -2,20 +2,18 @@
 
 require 'test_helper'
 
-module Members
-  class SearchComponentTest < ViewComponent::TestCase
-    test 'Should render a searchbox for a table of members' do
-      tab = ''
-      search_attribute = :user_email_cont
-      placeholder = 'a placeholder'
-      url = '/-/groups/group-1/-/members'
-      render_inline SearchComponent.new(Member.ransack, tab, url, search_attribute, placeholder)
+class SearchComponentTest < ViewComponent::TestCase
+  test 'Should render a searchbox' do
+    tab = ''
+    search_attribute = :user_email_cont
+    placeholder = 'a placeholder'
+    url = '/-/groups/group-1/-/members'
+    render_inline SearchComponent.new(Member.ransack, tab, url, search_attribute, placeholder)
 
-      assert_selector "input[type='hidden'][name='tab'][value='#{tab}']", visible: false, count: 1
-      assert_selector "form[action='#{url}']", count: 1
-      assert_selector "label[for='q_#{search_attribute}']", count: 1
-      assert_selector "input[id='q_#{search_attribute}']", count: 1
-      assert_text placeholder
-    end
+    assert_selector "input[type='hidden'][name='tab'][value='#{tab}']", visible: false, count: 1
+    assert_selector "form[action='#{url}']", count: 1
+    assert_selector "label[for='q_#{search_attribute}']", count: 1
+    assert_selector "input[id='q_#{search_attribute}']", count: 1
+    assert_text placeholder
   end
 end
